Rename commitTypeSelect folder
